+++ Libplanet.Tests/Net/Protocols/TestTransport.cs
@@ -10,6 +10,7 @@ using Libplanet.Crypto;
 using Libplanet.Net;
 using Libplanet.Net.Messages;
 using Libplanet.Net.Protocols;
+using Libplanet.Net.Transports;
 using Nito.AsyncEx;
 using Serilog;