+++ CHANGES.md
@@ -9,6 +9,8 @@ To be released.
 ### Backward-incompatible API changes
 
  -  Removed `StunMessage.Parse(Stream)` method.  [[#1228]]
+ -  Moved `ITransport` and `NetMQTransport` from `Libplanet.Net` to
+    `Libplanet.Net.Transports`.  [[#1235]]
 
 ### Backward-incompatible network protocol changes
 
@@ -19,6 +21,7 @@ To be released.
  -  Added `StunMessage.ParseAsync(Stream, CancellationToken)` method.
     [[#1228]]
  -  Added `Swarm<T>.AddPeersAsync()` method.  [[#1234]]
+ -  Added `NetMQTransport.QueryAppProtocolVersion()` static method.  [[#1235]]
 
 ### Behavioral changes
 
@@ -36,6 +39,7 @@ To be released.
 [#1219]: https://github.com/planetarium/libplanet/pull/1219
 [#1228]: https://github.com/planetarium/libplanet/pull/1218
 [#1234]: https://github.com/planetarium/libplanet/pull/1234
+[#1235]: https://github.com/planetarium/libplanet/pull/1235
 
 
 Version 0.11.0